+++ readme.md
@@ -1,0 +1,27 @@
+# pickmyrec-downloader
+
+> This readme is automatically generated by [create-ink-app](https://github.com/vadimdemedes/create-ink-app)
+
+
+## Install
+
+```bash
+$ npm install --global pickmyrec-downloader
+```
+
+
+## CLI
+
+```
+$ pickmyrec-downloader --help
+
+  Usage
+    $ pickmyrec-downloader
+
+  Options
+    --name  Your name
+
+  Examples
+    $ pickmyrec-downloader --name=Jane
+    Hello, Jane
+```
