+++ tsconfig.json
@@ -3,25 +3,28 @@
     "types": ["node"],
     "typeRoots": ["node_modules/@types"],
     "target": "es2017",
+    "outDir": "lib",
     "module": "commonjs",
     "moduleResolution": "node",
-    "allowJs": false,
+    "lib": ["es6", "dom"],
+    "baseUrl": ".",
+    "jsx": "react-native",
     "allowSyntheticDefaultImports": false,
     "noEmitOnError": true,
     "noImplicitReturns": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
-    "noImplicitAny": true,
+    "noImplicitAny": false,
     "noImplicitThis": true,
-    "alwaysStrict": true,
+    "alwaysStrict": false,
     "skipLibCheck": true,
     "strictNullChecks": true,
     "strictFunctionTypes": true,
     "strictPropertyInitialization": true
   },
   "typeAcquisition": {
-    "enable": false
+    "enable": true
   },
-  "include": ["global.ts", "./**/*"],
+  "include": ["./src/**/*"],
   "exclude": ["node_modules"]
 }
