+++ typings/ink-progress-bar.ts
@@ -1,0 +1,12 @@
+declare module "ink-progress-bar" {
+  interface ProgressBarProps {
+    character?: string;
+    percent?: number;
+    left?: number;
+    right?: number;
+  }
+  export default class ProgressBar extends React.Component<
+    ProgressBarProps,
+    any
+  > {}
+}
